ENH: change the approach for populating segment content in ITK image
Instead of assuming that all frames covering volume are present, paste individual
frames into the appropriate location in the reconstructed image geometry.

Conversion for David's SPL atlas object is now working, although the segmentation
does not line up with the greyscale image, probably due to the incorrect handling
of PixelSpacing in DCMTK ITK reader [1].

Open issue: it seems that a segment that has "gaps" (empty frames not included in
the segment) is a valid object. It is also perfectly fine to have a segment consisting
of a single frame. However, SliceThickness does not seem to be mandatory in the
PixelMeasuresSequence. Need to confirm with David.

[1] https://github.com/InsightSoftwareConsortium/ITK/blob/master/Modules/IO/DCMTK/src/itkDCMTKFileReader.cxx#L1322-L1328